pojo类修改,增加Result表示返回结果

+++ src/main/java/com/dhu/swimmingpool/Pojo/Card.java
@@ -14,5 +14,5 @@ public class Card {
   private double balance;
   private long tid;
   private double discount;
-  private long status;
+  private int status;
 }
